requirements.txt
pyyaml
pydantic
sqlalchemy
aiohttp
asyncio
typer[all]
openai
anthropic
pandas
jupyter
matplotlib
seaborn